Move RawWrite and RawReadResponse out of crucible_protocol
