+++ downstairs/src/lib.rs
@@ -21,8 +21,8 @@ use crucible_common::{
 };
 use crucible_protocol::{
     BlockContext, CrucibleDecoder, CrucibleEncoder, JobId, Message,
-    ReadRequest, ReadResponse, ReconciliationId, SnapshotDetails,
-    CRUCIBLE_MESSAGE_VERSION,
+    ReadRequest, ReadResponse, ReadResponseBlockMetadata, ReconciliationId,
+    SnapshotDetails, CRUCIBLE_MESSAGE_VERSION,
 };
 use repair_client::Client;
 
@@ -118,6 +118,48 @@ impl IOop {
             | IOop::ExtentLiveReopen { dependencies, .. }
             | IOop::ExtentLiveNoOp { dependencies } => dependencies,
         }
+    }
+}
+
+/// Read response data, containing data from all blocks
+#[derive(Debug)]
+pub struct RawReadResponse {
+    /// Per-block metadata
+    pub blocks: Vec<ReadResponseBlockMetadata>,
+    /// Raw data
+    pub data: bytes::BytesMut,
+}
+
+impl RawReadResponse {
+    /// Builds a new empty `RawReadResponse` with the given capacity
+    pub fn with_capacity(block_count: usize, block_size: u64) -> Self {
+        Self {
+            blocks: Vec::with_capacity(block_count),
+            data: bytes::BytesMut::with_capacity(
+                block_count * block_size as usize,
+            ),
+        }
+    }
+
+    /// Destructures into a `Vec<ReadResponse>`
+    ///
+    /// This is useful for backwards compatibility in unit tests
+    #[cfg(test)]
+    pub fn into_read_responses(mut self) -> Vec<ReadResponse> {
+        assert_eq!(self.data.len() % self.blocks.len(), 0);
+        let block_size = self.data.len() / self.blocks.len();
+        let mut out = Vec::with_capacity(self.blocks.len());
+        for b in self.blocks {
+            let data = self.data.split_to(block_size);
+            out.push(ReadResponse {
+                eid: b.eid,
+                offset: b.offset,
+                block_contexts: b.block_contexts,
+                data,
+            })
+        }
+        assert!(self.data.is_empty());
+        out
     }
 }
